requirements.txt
# Django
Django>=5.2,<6.0

# Database
# psycopg2-binary>=2.9  # PostgreSQL için gerekirse ekle

# Media
pillow>=11.0

# Rich Text Editor
django-ckeditor>=6.7

# Production Server
gunicorn>=23.0

# Environment
python-dotenv>=1.0
# Slug generation with Unicode support
python-slugify>=8.0